requirements.txt
aniso8601==9.0.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-JWT==0.3.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
inflect==5.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
mysql-connector-python==8.0.27
protobuf==3.19.1
PyJWT==1.4.2
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlacodegen==2.3.0
SQLAlchemy==1.4.10
Werkzeug==1.0.1
